implementacion filterUpWork
