+++ src/domains/dashBoard/activity/list/core/uses-cases/filterUpWork.js
@@ -1,0 +1,10 @@
+const filterUpWord = (activity, work) => {
+  const a = activity.filter(item => {
+    return item.description.includes(work)
+  })
+  debugger
+  console.log(a)
+  return a
+}
+
+export default filterUpWord
