+++ src/domains/dashBoard/activity/list/ui/components/filterForm/src/FilterForm.jsx
@@ -1,16 +1,23 @@
 'use client'
 import '../styles/main.scss'
 import Image from 'next/image'
-import {useState} from 'react'
+import {useState, useEffect} from 'react'
 
 import InputSearch from '@domains/dashBoard/home/ui/components/inputSearch'
 import CardActivity from '@domains/cross/ui/components/cardActivity/src/CardActivity'
 import ListActivity from '@domains/dashBoard/home/ui/components/listActivity/src/ListActivity'
 import ModalFilter from '@domains/dashBoard/activity/list/ui/components/modalFilter/src/ModalFilter'
 import useActivity from '@domains/dashBoard/home/core/hooks/useActivity/src/useActivity'
+import filterUpWord from '@domains/dashBoard/activity/list/core/uses-cases/filterUpWork'
 
 const FilterForm = ({slug, showButton}) => {
-  const {activity} = useActivity({searchItem: slug})
+  const {activity, updateStateActivity} = useActivity({searchItem: slug})
+
+  const [filterActivity, setFilterActivity] = useState([])
+
+  useEffect(() => {
+    setFilterActivity(filterUpWord(activity, 'Deposito'))
+  }, [activity])
 
   const [isModalOpen, setIsModalOpen] = useState(false)
 
@@ -18,7 +25,7 @@ const FilterForm = ({slug, showButton}) => {
     setIsModalOpen(!isModalOpen)
   }
   const customClass = 'filter-form'
-
+  debugger
   return (
     <div className={customClass}>
       <div className={`${customClass}__form`}>
@@ -45,8 +52,8 @@ const FilterForm = ({slug, showButton}) => {
       </div>
       <div className="filter-form__activity">
         <CardActivity size="Activity">
-          {activity.length > 0 ? (
-            <ListActivity activity={activity} />
+          {filterActivity.length > 0 ? (
+            <ListActivity activity={filterActivity} />
           ) : (
             <p>No hay actividades disponibles</p>
           )}
