+++ src/domains/dashBoard/home/ui/composite-wrappers/CWactivity/src/CWActivity.jsx
@@ -8,12 +8,6 @@ import useActivity from '@domains/dashBoard/home/core/hooks/useActivity/src/useA
 
 const CWActivity = () => {
   const defaultClass = 'activity-money-component'
-  const {activity} = useActivity()
-  debugger
-
-  useEffect(() => {
-    console.log(activity)
-  }, [activity])
 
   return (
     <>
